Added es6 for plugin.resume.js
