+++ .gitattributes
@@ -3,3 +3,4 @@ BookReader/plugins/*.js.map binary
 BookReader/plugins/plugin.archive_analytics.js binary
 BookReader/plugins/plugin.menu_toggle.js binary
 BookReader/plugins/plugin.tts.js binary
+BookReader/plugins/plugin.resume.js binary
