+++ BookReader/plugins/plugin.resume.js
@@ -1,73 +1,2 @@
-/**
- * Plugin to remember the current page in a cookie
- */
-
-jQuery.extend(BookReader.defaultOptions, {
-    enablePageResume: true,
-    /**
-     * @var null|string eg '/', '/details/id'
-     */
-    resumeCookiePath: null,
-});
-
-BookReader.prototype.init = (function(super_) {
-    return function() {
-        super_.call(this);
-        if (this.options.enablePageResume) {
-            this.bind(BookReader.eventNames.fragmentChange, function() {
-                var params = this.paramsFromCurrent();
-                this.updateResumeValue(params.index);
-            }.bind(this));
-        }
-    };
-})(BookReader.prototype.init);
-
-/**
- * Get's the page resume value, for remembering reader's page
- * Can be overriden for different implementation
- * @return {Number|null}
- */
-BookReader.prototype.getResumeValue = function() {
-    var val = BookReader.docCookies.getItem('br-resume');
-    if (val !== null) return parseInt(val);
-    else return null;
-}
-
-/**
- * Set's the page resume value, for remembering reader's page
- * Can be overriden for different implementation
- * @param {Number} the leaf index
- */
-BookReader.prototype.updateResumeValue = function(index, cookieName) {
-    var ttl = new Date(+new Date + 12096e5); // 2 weeks
-    var path = this.options.resumeCookiePath || window.location.pathname;
-    cookieName = cookieName || 'br-resume';
-    BookReader.docCookies.setItem(cookieName, index, ttl, path, null, false);
-}
-
-/*\
-|*|  https://developer.mozilla.org/en-US/docs/Web/API/document.cookie
-|*|  https://developer.mozilla.org/User:fusionchess
-|*|  https://github.com/madmurphy/cookies.js
-|*|  This framework is released under the GNU Public License, version 3 or later.
-|*|  http://www.gnu.org/licenses/gpl-3.0-standalone.html
-\*/
-BookReader.docCookies = {
-  getItem: function (sKey) {
-    if (!sKey) { return null; }
-    return decodeURIComponent(document.cookie.replace(new RegExp("(?:(?:^|.*;)\\s*" + encodeURIComponent(sKey).replace(/[\-\.\+\*]/g, "\\$&") + "\\s*\\=\\s*([^;]*).*$)|^.*$"), "$1")) || null;
-  },
-  setItem: function (sKey, sValue, vEnd, sPath, sDomain, bSecure) {
-    var sExpires = "";
-    if (vEnd) {
-      sExpires = "; expires=" + vEnd.toUTCString();
-    }
-    document.cookie = encodeURIComponent(sKey) + "=" + encodeURIComponent(sValue) + sExpires + (sDomain ? "; domain=" + sDomain : "") + (sPath ? "; path=" + sPath : "") + (bSecure ? "; secure" : "");
-    return true;
-  },
-  removeItem: function (sKey, sPath, sDomain) {
-    if (!this.hasItem(sKey)) { return false; }
-    document.cookie = encodeURIComponent(sKey) + "=; expires=Thu, 01 Jan 1970 00:00:00 GMT" + (sDomain ? "; domain=" + sDomain : "") + (sPath ? "; path=" + sPath : "");
-    return true;
-  },
-};
+!function(t){var n={};function e(r){if(n[r])return n[r].exports;var o=n[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,e),o.l=!0,o.exports}e.m=t,e.c=n,e.d=function(t,n,r){e.o(t,n)||Object.defineProperty(t,n,{enumerable:!0,get:r})},e.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},e.t=function(t,n){if(1&n&&(t=e(t)),8&n)return t;if(4&n&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(e.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&n&&"string"!=typeof t)for(var o in t)e.d(r,o,function(n){return t[n]}.bind(null,o));return r},e.n=function(t){var n=t&&t.__esModule?function(){return t.default}:function(){return t};return e.d(n,"a",n),n},e.o=function(t,n){return Object.prototype.hasOwnProperty.call(t,n)},e.p="",e(e.s=22)}([function(t,n,e){(function(n){var e=function(t){return t&&t.Math==Math&&t};t.exports=e("object"==typeof globalThis&&globalThis)||e("object"==typeof window&&window)||e("object"==typeof self&&self)||e("object"==typeof n&&n)||Function("return this")()}).call(this,e(24))},function(t,n){t.exports=function(t){try{return!!t()}catch(t){return!0}}},function(t,n){t.exports=function(t){return"object"==typeof t?null!==t:"function"==typeof t}},function(t,n,e){var r=e(1);t.exports=!r((function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a}))},function(t,n,e){var r=e(2);t.exports=function(t){if(!r(t))throw TypeError(String(t)+" is not an object");return t}},function(t,n){t.exports=function(t){if(null==t)throw TypeError("Can't call method on "+t);return t}},function(t,n){var e=Math.ceil,r=Math.floor;t.exports=function(t){return isNaN(t=+t)?0:(t>0?r:e)(t)}},function(t,n,e){var r=e(0),o=e(8),i=e(19),u=e(42),c=r.Symbol,a=o("wks");t.exports=function(t){return a[t]||(a[t]=u&&c[t]||(u?c:i)("Symbol."+t))}},function(t,n,e){var r=e(39),o=e(40);(t.exports=function(t,n){return o[t]||(o[t]=void 0!==n?n:{})})("versions",[]).push({version:"3.3.3",mode:r?"pure":"global",copyright:"© 2019 Denis Pushkarev (zloirock.ru)"})},function(t,n,e){var r=e(3),o=e(10),i=e(41);t.exports=r?function(t,n,e){return o.f(t,n,i(1,e))}:function(t,n,e){return t[n]=e,t}},function(t,n,e){var r=e(3),o=e(29),i=e(4),u=e(31),c=Object.defineProperty;n.f=r?c:function(t,n,e){if(i(t),n=u(n,!0),i(e),o)try{return c(t,n,e)}catch(t){}if("get"in e||"set"in e)throw TypeError("Accessors not supported");return"value"in e&&(t[n]=e.value),t}},function(t,n){var e={}.hasOwnProperty;t.exports=function(t,n){return e.call(t,n)}},function(t,n){var e={}.toString;t.exports=function(t){return e.call(t).slice(8,-1)}},function(t,n,e){"use strict";var r=e(4);t.exports=function(){var t=r(this),n="";return t.global&&(n+="g"),t.ignoreCase&&(n+="i"),t.multiline&&(n+="m"),t.dotAll&&(n+="s"),t.unicode&&(n+="u"),t.sticky&&(n+="y"),n}},function(t,n,e){var r=e(0),o=e(8),i=e(9),u=e(11),c=e(18),a=e(20),f=e(43),s=f.get,l=f.enforce,p=String(a).split("toString");o("inspectSource",(function(t){return a.call(t)})),(t.exports=function(t,n,e,o){var a=!!o&&!!o.unsafe,f=!!o&&!!o.enumerable,s=!!o&&!!o.noTargetGet;"function"==typeof e&&("string"!=typeof n||u(e,"name")||i(e,"name",n),l(e).source=p.join("string"==typeof n?n:"")),t!==r?(a?!s&&t[n]&&(f=!0):delete t[n],f?t[n]=e:i(t,n,e)):f?t[n]=e:c(n,e)})(Function.prototype,"toString",(function(){return"function"==typeof this&&s(this).source||a.call(this)}))},function(t,n,e){var r=e(34),o=e(5);t.exports=function(t){return r(o(t))}},function(t,n,e){var r=e(6),o=Math.min;t.exports=function(t){return t>0?o(r(t),9007199254740991):0}},function(t,n){t.exports={}},function(t,n,e){var r=e(0),o=e(9);t.exports=function(t,n){try{o(r,t,n)}catch(e){r[t]=n}return n}},function(t,n){var e=0,r=Math.random();t.exports=function(t){return"Symbol("+String(void 0===t?"":t)+")_"+(++e+r).toString(36)}},function(t,n,e){var r=e(8);t.exports=r("native-function-to-string",Function.toString)},function(t,n,e){"use strict";var r,o,i=e(13),u=RegExp.prototype.exec,c=String.prototype.replace,a=u,f=(r=/a/,o=/b*/g,u.call(r,"a"),u.call(o,"a"),0!==r.lastIndex||0!==o.lastIndex),s=void 0!==/()??/.exec("")[1];(f||s)&&(a=function(t){var n,e,r,o,a=this;return s&&(e=new RegExp("^"+a.source+"$(?!\\s)",i.call(a))),f&&(n=a.lastIndex),r=u.call(a,t),f&&r&&(a.lastIndex=a.global?r.index+r[0].length:n),s&&r&&r.length>1&&c.call(r[0],e,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(r[o]=void 0)})),r}),t.exports=a},function(t,n,e){"use strict";e.r(n);var r;e(23),e(49),e(50);jQuery.extend(BookReader.defaultOptions,{enablePageResume:!0,resumeCookiePath:null}),BookReader.prototype.init=(r=BookReader.prototype.init,function(){r.call(this),this.options.enablePageResume&&this.bind(BookReader.eventNames.fragmentChange,function(){var t=this.paramsFromCurrent();this.updateResumeValue(t.index)}.bind(this))}),BookReader.prototype.getResumeValue=function(){var t=BookReader.docCookies.getItem("br-resume");return null!==t?parseInt(t):null},BookReader.prototype.updateResumeValue=function(t,n){var e=new Date(+new Date+12096e5),r=this.options.resumeCookiePath||window.location.pathname;n=n||"br-resume",BookReader.docCookies.setItem(n,t,e,r,null,!1)},BookReader.docCookies={getItem:function(t){return t&&decodeURIComponent(document.cookie.replace(new RegExp("(?:(?:^|.*;)\\s*"+encodeURIComponent(t).replace(/[\-\.\+\*]/g,"\\$&")+"\\s*\\=\\s*([^;]*).*$)|^.*$"),"$1"))||null},setItem:function(t,n,e,r,o,i){var u="";return e&&(u="; expires="+e.toUTCString()),document.cookie=encodeURIComponent(t)+"="+encodeURIComponent(n)+u+(o?"; domain="+o:"")+(r?"; path="+r:"")+(i?"; secure":""),!0},removeItem:function(t,n,e){return!!this.hasItem(t)&&(document.cookie=encodeURIComponent(t)+"=; expires=Thu, 01 Jan 1970 00:00:00 GMT"+(e?"; domain="+e:"")+(n?"; path="+n:""),!0)}}},function(t,n,e){var r=e(3),o=e(0),i=e(25),u=e(26),c=e(10).f,a=e(32).f,f=e(38),s=e(13),l=e(14),p=e(1),v=e(46),g=e(7)("match"),d=o.RegExp,h=d.prototype,x=/a/g,y=/a/g,m=new d(x)!==x;if(r&&i("RegExp",!m||p((function(){return y[g]=!1,d(x)!=x||d(y)==y||"/a/i"!=d(x,"i")})))){for(var b=function(t,n){var e=this instanceof b,r=f(t),o=void 0===n;return!e&&r&&t.constructor===b&&o?t:u(m?new d(r&&!o?t.source:t,n):d((r=t instanceof b)?t.source:t,r&&o?s.call(t):n),e?this:h,b)},S=function(t){t in b||c(b,t,{configurable:!0,get:function(){return d[t]},set:function(n){d[t]=n}})},w=a(d),R=0;w.length>R;)S(w[R++]);h.constructor=b,b.prototype=h,l(o,"RegExp",b)}v("RegExp")},function(t,n){var e;e=function(){return this}();try{e=e||new Function("return this")()}catch(t){"object"==typeof window&&(e=window)}t.exports=e},function(t,n,e){var r=e(1),o=/#|\.prototype\./,i=function(t,n){var e=c[u(t)];return e==f||e!=a&&("function"==typeof n?r(n):!!n)},u=i.normalize=function(t){return String(t).replace(o,".").toLowerCase()},c=i.data={},a=i.NATIVE="N",f=i.POLYFILL="P";t.exports=i},function(t,n,e){var r=e(2),o=e(27);t.exports=function(t,n,e){var i,u;return o&&"function"==typeof(i=n.constructor)&&i!==e&&r(u=i.prototype)&&u!==e.prototype&&o(t,u),t}},function(t,n,e){var r=e(4),o=e(28);t.exports=Object.setPrototypeOf||("__proto__"in{}?function(){var t,n=!1,e={};try{(t=Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set).call(e,[]),n=e instanceof Array}catch(t){}return function(e,i){return r(e),o(i),n?t.call(e,i):e.__proto__=i,e}}():void 0)},function(t,n,e){var r=e(2);t.exports=function(t){if(!r(t)&&null!==t)throw TypeError("Can't set "+String(t)+" as a prototype");return t}},function(t,n,e){var r=e(3),o=e(1),i=e(30);t.exports=!r&&!o((function(){return 7!=Object.defineProperty(i("div"),"a",{get:function(){return 7}}).a}))},function(t,n,e){var r=e(0),o=e(2),i=r.document,u=o(i)&&o(i.createElement);t.exports=function(t){return u?i.createElement(t):{}}},function(t,n,e){var r=e(2);t.exports=function(t,n){if(!r(t))return t;var e,o;if(n&&"function"==typeof(e=t.toString)&&!r(o=e.call(t)))return o;if("function"==typeof(e=t.valueOf)&&!r(o=e.call(t)))return o;if(!n&&"function"==typeof(e=t.toString)&&!r(o=e.call(t)))return o;throw TypeError("Can't convert object to primitive value")}},function(t,n,e){var r=e(33),o=e(37).concat("length","prototype");n.f=Object.getOwnPropertyNames||function(t){return r(t,o)}},function(t,n,e){var r=e(11),o=e(15),i=e(35).indexOf,u=e(17);t.exports=function(t,n){var e,c=o(t),a=0,f=[];for(e in c)!r(u,e)&&r(c,e)&&f.push(e);for(;n.length>a;)r(c,e=n[a++])&&(~i(f,e)||f.push(e));return f}},function(t,n,e){var r=e(1),o=e(12),i="".split;t.exports=r((function(){return!Object("z").propertyIsEnumerable(0)}))?function(t){return"String"==o(t)?i.call(t,""):Object(t)}:Object},function(t,n,e){var r=e(15),o=e(16),i=e(36),u=function(t){return function(n,e,u){var c,a=r(n),f=o(a.length),s=i(u,f);if(t&&e!=e){for(;f>s;)if((c=a[s++])!=c)return!0}else for(;f>s;s++)if((t||s in a)&&a[s]===e)return t||s||0;return!t&&-1}};t.exports={includes:u(!0),indexOf:u(!1)}},function(t,n,e){var r=e(6),o=Math.max,i=Math.min;t.exports=function(t,n){var e=r(t);return e<0?o(e+n,0):i(e,n)}},function(t,n){t.exports=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"]},function(t,n,e){var r=e(2),o=e(12),i=e(7)("match");t.exports=function(t){var n;return r(t)&&(void 0!==(n=t[i])?!!n:"RegExp"==o(t))}},function(t,n){t.exports=!1},function(t,n,e){var r=e(0),o=e(18),i=r["__core-js_shared__"]||o("__core-js_shared__",{});t.exports=i},function(t,n){t.exports=function(t,n){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:n}}},function(t,n,e){var r=e(1);t.exports=!!Object.getOwnPropertySymbols&&!r((function(){return!String(Symbol())}))},function(t,n,e){var r,o,i,u=e(44),c=e(0),a=e(2),f=e(9),s=e(11),l=e(45),p=e(17),v=c.WeakMap;if(u){var g=new v,d=g.get,h=g.has,x=g.set;r=function(t,n){return x.call(g,t,n),n},o=function(t){return d.call(g,t)||{}},i=function(t){return h.call(g,t)}}else{var y=l("state");p[y]=!0,r=function(t,n){return f(t,y,n),n},o=function(t){return s(t,y)?t[y]:{}},i=function(t){return s(t,y)}}t.exports={set:r,get:o,has:i,enforce:function(t){return i(t)?o(t):r(t,{})},getterFor:function(t){return function(n){var e;if(!a(n)||(e=o(n)).type!==t)throw TypeError("Incompatible receiver, "+t+" required");return e}}}},function(t,n,e){var r=e(0),o=e(20),i=r.WeakMap;t.exports="function"==typeof i&&/native code/.test(o.call(i))},function(t,n,e){var r=e(8),o=e(19),i=r("keys");t.exports=function(t){return i[t]||(i[t]=o(t))}},function(t,n,e){"use strict";var r=e(47),o=e(10),i=e(7),u=e(3),c=i("species");t.exports=function(t){var n=r(t),e=o.f;u&&n&&!n[c]&&e(n,c,{configurable:!0,get:function(){return this}})}},function(t,n,e){var r=e(48),o=e(0),i=function(t){return"function"==typeof t?t:void 0};t.exports=function(t,n){return arguments.length<2?i(r[t])||i(o[t]):r[t]&&r[t][n]||o[t]&&o[t][n]}},function(t,n,e){t.exports=e(0)},function(t,n,e){"use strict";var r=e(14),o=e(4),i=e(1),u=e(13),c=RegExp.prototype,a=c.toString,f=i((function(){return"/a/b"!=a.call({source:"a",flags:"b"})})),s="toString"!=a.name;(f||s)&&r(RegExp.prototype,"toString",(function(){var t=o(this),n=String(t.source),e=t.flags;return"/"+n+"/"+String(void 0===e&&t instanceof RegExp&&!("flags"in c)?u.call(t):e)}),{unsafe:!0})},function(t,n,e){"use strict";var r=e(51),o=e(4),i=e(52),u=e(16),c=e(6),a=e(5),f=e(53),s=e(55),l=Math.max,p=Math.min,v=Math.floor,g=/\$([$&'`]|\d\d?|<[^>]*>)/g,d=/\$([$&'`]|\d\d?)/g;r("replace",2,(function(t,n,e){return[function(e,r){var o=a(this),i=null==e?void 0:e[t];return void 0!==i?i.call(e,o,r):n.call(String(o),e,r)},function(t,i){var a=e(n,t,this,i);if(a.done)return a.value;var v=o(t),g=String(this),d="function"==typeof i;d||(i=String(i));var h=v.global;if(h){var x=v.unicode;v.lastIndex=0}for(var y=[];;){var m=s(v,g);if(null===m)break;if(y.push(m),!h)break;""===String(m[0])&&(v.lastIndex=f(g,u(v.lastIndex),x))}for(var b,S="",w=0,R=0;R<y.length;R++){m=y[R];for(var j=String(m[0]),O=l(p(c(m.index),g.length),0),E=[],_=1;_<m.length;_++)E.push(void 0===(b=m[_])?b:String(b));var k=m.groups;if(d){var I=[j].concat(E,O,g);void 0!==k&&I.push(k);var P=String(i.apply(void 0,I))}else P=r(j,g,O,E,k,i);O>=w&&(S+=g.slice(w,O)+P,w=O+j.length)}return S+g.slice(w)}];function r(t,e,r,o,u,c){var a=r+t.length,f=o.length,s=d;return void 0!==u&&(u=i(u),s=g),n.call(c,s,(function(n,i){var c;switch(i.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,r);case"'":return e.slice(a);case"<":c=u[i.slice(1,-1)];break;default:var s=+i;if(0===s)return n;if(s>f){var l=v(s/10);return 0===l?n:l<=f?void 0===o[l-1]?i.charAt(1):o[l-1]+i.charAt(1):n}c=o[s-1]}return void 0===c?"":c}))}}))},function(t,n,e){"use strict";var r=e(9),o=e(14),i=e(1),u=e(7),c=e(21),a=u("species"),f=!i((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),s=!i((function(){var t=/(?:)/,n=t.exec;t.exec=function(){return n.apply(this,arguments)};var e="ab".split(t);return 2!==e.length||"a"!==e[0]||"b"!==e[1]}));t.exports=function(t,n,e,l){var p=u(t),v=!i((function(){var n={};return n[p]=function(){return 7},7!=""[t](n)})),g=v&&!i((function(){var n=!1,e=/a/;return e.exec=function(){return n=!0,null},"split"===t&&(e.constructor={},e.constructor[a]=function(){return e}),e[p](""),!n}));if(!v||!g||"replace"===t&&!f||"split"===t&&!s){var d=/./[p],h=e(p,""[t],(function(t,n,e,r,o){return n.exec===c?v&&!o?{done:!0,value:d.call(n,e,r)}:{done:!0,value:t.call(e,n,r)}:{done:!1}})),x=h[0],y=h[1];o(String.prototype,t,x),o(RegExp.prototype,p,2==n?function(t,n){return y.call(t,this,n)}:function(t){return y.call(t,this)}),l&&r(RegExp.prototype[p],"sham",!0)}}},function(t,n,e){var r=e(5);t.exports=function(t){return Object(r(t))}},function(t,n,e){"use strict";var r=e(54).charAt;t.exports=function(t,n,e){return n+(e?r(t,n).length:1)}},function(t,n,e){var r=e(6),o=e(5),i=function(t){return function(n,e){var i,u,c=String(o(n)),a=r(e),f=c.length;return a<0||a>=f?t?"":void 0:(i=c.charCodeAt(a))<55296||i>56319||a+1===f||(u=c.charCodeAt(a+1))<56320||u>57343?t?c.charAt(a):i:t?c.slice(a,a+2):u-56320+(i-55296<<10)+65536}};t.exports={codeAt:i(!1),charAt:i(!0)}},function(t,n,e){var r=e(12),o=e(21);t.exports=function(t,n){var e=t.exec;if("function"==typeof e){var i=e.call(t,n);if("object"!=typeof i)throw TypeError("RegExp exec method returned something other than an Object or null");return i}if("RegExp"!==r(t))throw TypeError("RegExp#exec called on incompatible receiver");return o.call(t,n)}}]);
+//# sourceMappingURL=plugin.resume.js.map
